+++ pypalex/Extractor.py
@@ -5,6 +5,7 @@
 #   - Created by Al Timofeyev on February 10, 2022.
 #   - Modified by Al Timofeyev on April 21, 2022.
 #   - Modified by Al Timofeyev on March 6, 2023.
+#   - Modified by Al Timofeyev on March 22, 2023.
 
 
 # ---- IMPORTS ----
@@ -21,21 +22,23 @@ class Extractor:
     #   @param  self                The object pointer.
     #   @param  hsv_img_matrix_2d   A 2D numpy array of pixels from an image in [h,s,v] format.
     #   @param  output_filepath     Output file path with filename of where to store color palette.
-    #   @param  pastel              Flag to convert all extracted color types to pastel.
-    #   @param  pastel_light        Flag to convert light colors to pastel.
-    #   @param  pastel_normal       Flag to convert normal colors to pastel.
-    #   @param  pastel_dark         Flag to convert dark colors to pastel.
-    def __init__(self, hsv_img_matrix_2d, output_filepath, pastel=False, pastel_light=False, pastel_normal=False, pastel_dark=False):
+    #   @param  pastel_light        Flag to convert light color palette to pastel.
+    #   @param  pastel_normal       Flag to convert normal color palette to pastel.
+    #   @param  pastel_dark         Flag to convert dark color palette to pastel.
+    #   @param  sat_pref_light      Flag that gives preference to more saturated colors of the light color palette.
+    #   @param  sat_pref_normal     Flag that gives preference to more saturated colors of the normal color palette.
+    #   @param  sat_pref_dark       Flag that gives preference to more saturated colors of the dark color palette.
+    def __init__(self, hsv_img_matrix_2d, output_filepath, pastel_light=False, pastel_normal=False, pastel_dark=False, sat_pref_light=False, sat_pref_normal=False, sat_pref_dark=False):
         self.hsv_img_matrix_2d = hsv_img_matrix_2d
         self.output_filepath = output_filepath
-        self.pastel = pastel
         self.pastel_light = pastel_light
         self.pastel_normal = pastel_normal
         self.pastel_dark = pastel_dark
+        self.sat_pref_list = [sat_pref_light, sat_pref_normal, sat_pref_dark]
         self.ratio_dict = {}
         self.base_color_dict = {}
         self.extracted_colors_dict = {}
-        self.color_palette_dict = {}
+        self.color_schemes_dict = {}
 
     # --------------------------------------------------------------------------
     # --------------------------------------------------------------------------
@@ -50,15 +53,15 @@ class Extractor:
         self.base_color_dict = exutil.construct_base_color_dictionary(self.hsv_img_matrix_2d)
 
         # Extract colors.
-        self.extracted_colors_dict = exutil.extract_colors(self.base_color_dict)
+        self.extracted_colors_dict = exutil.extract_color_palettes(self.base_color_dict, self.sat_pref_list)
         exutil.check_missing_colors(self.base_color_dict, self.extracted_colors_dict)
         exutil.generate_remaining_colors(self.extracted_colors_dict, self.ratio_dict)
 
         # Check for pastel option(s).
         self.check_pastel_conversion()
 
-        # Construct palette dictionary.
-        self.construct_palette_dictionary()
+        # Construct color scheme dictionary with a mix of all 3 color palettes.
+        self.construct_scheme_dictionary()
 
     # **************************************************************************
     # **************************************************************************
@@ -67,12 +70,6 @@ class Extractor:
     #
     #   @param  self    The object pointer.
     def check_pastel_conversion(self):
-        if self.pastel:
-            self.convert_pastel_light()
-            self.convert_pastel_normal()
-            self.convert_pastel_dark()
-            return
-
         if self.pastel_light:
             self.convert_pastel_light()
 
@@ -85,11 +82,14 @@ class Extractor:
     # --------------------------------------------------------------------------
     # --------------------------------------------------------------------------
 
-    ##  Constructs color palette dictionary.
+    ##  Constructs a dictionary of color schemes by combining color palettes.
+    #   @details    Light color scheme contains the normal and dark
+    #               color palettes. Dark color scheme contains the
+    #               normal and light color palettes.
     #
     #   @param  self    The object pointer.
-    def construct_palette_dictionary(self):
-        light_palette = {
+    def construct_scheme_dictionary(self):
+        light_scheme = {
             'background': convert.hsv_to_hex(self.extracted_colors_dict['Light Background']),
             'foreground': convert.hsv_to_hex(self.extracted_colors_dict['Dark Foreground']),
             'black': convert.hsv_to_hex(self.extracted_colors_dict['Normal Black']),
@@ -110,7 +110,7 @@ class Extractor:
             'bright white': convert.hsv_to_hex(self.extracted_colors_dict['Dark White'])
         }
 
-        dark_palette = {
+        dark_scheme = {
             'background': convert.hsv_to_hex(self.extracted_colors_dict['Dark Background']),
             'foreground': convert.hsv_to_hex(self.extracted_colors_dict['Light Foreground']),
             'black': convert.hsv_to_hex(self.extracted_colors_dict['Normal Black']),
@@ -131,8 +131,8 @@ class Extractor:
             'bright white': convert.hsv_to_hex(self.extracted_colors_dict['Light White'])
         }
 
-        self.color_palette_dict['light'] = light_palette
-        self.color_palette_dict['dark'] = dark_palette
+        self.color_schemes_dict['light'] = light_scheme
+        self.color_schemes_dict['dark'] = dark_scheme
 
     # **************************************************************************
     # **************************************************************************
@@ -208,14 +208,14 @@ class Extractor:
     #   A 2D numpy array of pixels from an image in [h,s,v] format.
     ##  @var    output_filepath
     #   Output file path with filename of where to store color palette.
-    ##  @var    pastel
-    #   Flag to convert all extracted color types to pastel.
     ##  @var    pastel_light
-    #   Flag to convert light colors to pastel.
+    #   Flag to convert light color palette to pastel.
     ##  @var    pastel_normal
-    #   Flag to convert normal colors to pastel.
+    #   Flag to convert normal color palette to pastel.
     ##  @var    pastel_dark
-    #   Flag to convert dark colors to pastel.
+    #   Flag to convert dark color palette to pastel.
+    ##  @var    sat_pref_list
+    #   List of saturation preference flags for light, normal, dark color palettes.
     ##  @var    ratio_dict
     #   A dictionary that holds the ratio of base colors in an image and is
     #   used to identify the dominant color in an image.
@@ -223,6 +223,6 @@ class Extractor:
     #   A dictionary of 2D numpy arrays for each of the 6 base colors.
     ##  @var    extracted_colors_dict
     #   A Dictionary of extracted colors in [h,s,v] format.
-    ##  @var    color_palette_dict
+    ##  @var    color_schemes_dict
     #   A Dictionary of dictionaries for light and dark color schemes
     #   that are in HEX string format.