+++ pypalex/settings.py
@@ -7,6 +7,7 @@
 #   - Created by Al Timofeyev on March 2, 2022
 #   - Modified by Al Timofeyev on April 21, 2022.
 #   - Modified by Al Timofeyev on March 11, 2023.
+#   - Modified by Al Timofeyev on March 22, 2023.
 #
 #   @note   This code has been borrowed from Dylan Araps
 #           PyWal on github: https://github.com/dylanaraps/pywal/blob/master/pywal/settings.py
@@ -15,7 +16,7 @@
 import os
 import platform
 
-__version__ = "1.3.0"
+__version__ = "1.3.1"
 __cache_version__ = "1.0.0"
 
 HOME = os.getenv("HOME", os.getenv("USERPROFILE"))